requirements.txt
python-telegram-bot==20.8
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
openpyxl==3.1.2
openai>=1.0.0
anthropic>=0.7.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
reportlab>=4.0.0
